pyproject.toml
[tool.poetry]
name = "blogpy"
version = "0.1.0"
description = ""
authors = ["RenegadeLoss <Renegadeloss696@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.1"
django-rest-framework = "^0.1.0"
psycopg2 = "^2.9.6"
pillow = "^9.5.0"
django-taggit = "^4.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
